feat(entities-shared): add useTableState
