+++ packages/entities/entities-shared/src/composables/index.ts
@@ -12,6 +12,7 @@ import useGatewayFeatureSupported from './useGatewayFeatureSupported'
 import useTruncationDetector from './useTruncationDetector'
 import useValidators from './useValidators'
 import { useSchemaProvider, useSubSchema } from './useSchema'
+import useTableState from './useTableState'
 
 // All composables must be exported as part of the default object for Cypress test stubs
 export default {
@@ -31,4 +32,5 @@ export default {
   useValidators,
   useSchemaProvider,
   useSubSchema,
+  useTableState,
 }
